create Utils function for device owner verification
